DAOS-7073 dtx: use latest pool map for DTX resync and refresh
There may be leader switch during the DTX resync and refresh.
If we used stale pool map, then we may find the wrong leader
that may be dead or does not know related DTX status.

If current server is the new DTX leader for DTX refresh, then
we need to do DTX check with other replicas as DTX resync does
to avoid blocking the refresh sponsor for very long time.

Signed-off-by: Fan Yong <fan.yong@intel.com>